Moved song player to a new component for now
