+++ client/src/app/home/home-components/song-card/song-card.component.scss
