+++ client/src/app/home/home-components/song-card/song-card.component.ts
@@ -1,5 +1,5 @@
 import { Component, Input } from '@angular/core';
-import { SongCardData } from '../../interfaces/SongCardData'; 
+import { SongCardData } from '../../../interfaces/SongCardData';
 
 @Component({
   selector: 'app-song-card',
@@ -8,6 +8,6 @@ import { SongCardData } from '../../interfaces/SongCardData';
   templateUrl: './song-card.component.html',
   styleUrl: './song-card.component.scss'
 })
-export class SongCardComponent{ 
+export class SongCardComponent{
   @Input() songCardData?: SongCardData;
 }
