+++ client/src/app/home/home-components/song-player/song-player.component.spec.ts
@@ -1,0 +1,23 @@
+import { ComponentFixture, TestBed } from '@angular/core/testing';
+
+import { SongPlayerComponent } from './song-player.component';
+
+describe('SongPlayerComponent', () => {
+  let component: SongPlayerComponent;
+  let fixture: ComponentFixture<SongPlayerComponent>;
+
+  beforeEach(async () => {
+    await TestBed.configureTestingModule({
+      imports: [SongPlayerComponent]
+    })
+    .compileComponents();
+    
+    fixture = TestBed.createComponent(SongPlayerComponent);
+    component = fixture.componentInstance;
+    fixture.detectChanges();
+  });
+
+  it('should create', () => {
+    expect(component).toBeTruthy();
+  });
+});
