+++ client/src/app/home/home.component.scss
@@ -1,24 +1,1 @@
-.content-wrapper {
-    width: 600px;
-    font-family: 'Segoe UI', sans-serif;
-}
-
-.player {
-    width: 100%;
-    background-color: rgb(35, 35, 35);
-    box-sizing: border-box;
-    padding: 20px;
-    border-radius: 20px;
-}
-
-.redirect-link {
-    text-decoration: none;
-    color: white;
-    transition: 150ms;
-}
-
-.redirect-link:hover {
-    color: rgb(181, 181, 181);
-}
-
