+++ client/src/app/login/login.component.ts
@@ -2,7 +2,7 @@ import { Component, inject } from '@angular/core';
 import { AuthService } from '../services/auth.service';
 
 @Component({
-  selector: 'login',
+  selector: 'app-login',
   standalone: true,
   imports: [],
   templateUrl: './login.component.html',
